+++ run/cli_runner.go
@@ -1,0 +1,28 @@
+package run
+
+import (
+	"github.com/nextmv-io/sdk/run/decode"
+	"github.com/nextmv-io/sdk/run/encode"
+)
+
+// CliRunner is the default CLI runner.
+func CliRunner[Input, Option, Solution any](
+	handler Algorithm[Input, Option, Solution],
+) Runner[Input, Option, Solution] {
+	runner := &genericRunner[Input, Option, Solution]{
+		IOProducer:    CliIOProducer,
+		InputDecoder:  GenericDecoder[Input, decode.JSONDecoder],
+		OptionDecoder: NoopOptionsDecoder[Option],
+		Algorithm:     handler,
+		Encoder:       GenericEncoder[Solution, Option, encode.JSONEncoder],
+	}
+	runnerConfig, decodedOption, err := FlagParser[
+		Option, CliRunnerConfig,
+	]()
+	runner.runnerConfig = runnerConfig
+	runner.decodedOption = decodedOption
+	if err != nil {
+		panic(err)
+	}
+	return runner
+}
